+++ clid/main.py
@@ -1,8 +1,11 @@
 #!/usr/bin/env python3
 
-__version__ = '0.5.3'
+__version__ = '0.6.0'
 
+import os
 import curses
+
+import configobj
 import npyscreen as npy
 
 from . import base
@@ -11,8 +14,9 @@ from . import database
 from . import editmeta
 
 
-class CommandLine(base.ClidCommandLine):
-    """Command line at the bottom.
+class MainActionController(base.ClidActionController):
+    """Object that recieves recieves inpout in command line
+       at the bottom.
 
        Note:
             self.parent refers to ClidInterface -> class
@@ -21,6 +25,16 @@ class CommandLine(base.ClidCommandLine):
     def create(self):
         super().create()
         self.add_action('^/.+', self.search_for_files, live=True)   # search with '/'
+
+    def change_setting(self, command_line, widget_proxy, live):
+        setting = command_line[5:].split(sep='=')
+        pref_form = self.parent.parentApp.getForm("SETTINGS")
+        pref_form.value.change_setting(setting[0], setting[1])   # writes to the ini file
+        pref_form.value.when_changed[setting[0]]()
+
+        pref_form.load_pref()
+        pref_form.wMain.display()
+
 
     def search_for_files(self, command_line, widget_proxy, live):
         """Search for files while given a string"""
@@ -122,18 +136,21 @@ class ClidInterface(npy.FormMuttActiveTraditional):
        Note:
             self.value refers to an instance of DATA_CONTROLER
     """
-    ACTION_CONTROLLER = CommandLine
-    MAIN_WIDGET_CLASS = ClidMultiline
     DATA_CONTROLER = database.Mp3DataBase
+    MAIN_WIDGET_CLASS = ClidMultiline
+    ACTION_CONTROLLER = MainActionController
+    COMMAND_WIDGET_CLASS = base.ClidCommandLine
 
-    def create(self):
-        super().create()   # so that status widgets are created
-        self.set_value(self.DATA_CONTROLER())   # isn't automatically done(open issue ?)
-
-        # used to revert screen(ESC) to standard view after a search(see class ClidMultiline)
-        self.after_search_now_filter_view = False
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.set_value(self.DATA_CONTROLER())
+        self.value.settings = self.parentApp.settings
+        self.value.load_files_and_set_values()
+        self.value.load_preview_format()
 
         self.load_files()
+
+        # widgets are created by self.create() in super()
         self.wStatus1.value = 'clid v' + __version__ + ' '
 
         try:
@@ -142,6 +159,9 @@ class ClidInterface(npy.FormMuttActiveTraditional):
             self.wStatus2.value = 'No Files Found In Directory'
             self.wMain.values = []
 
+        self.after_search_now_filter_view = False
+        # used to revert screen(ESC) to standard view after a search(see class ClidMultiline)
+
     # change to `load_pref`
     def load_files(self):
         """Set the mp3 files that will be displayed"""
@@ -149,10 +169,23 @@ class ClidInterface(npy.FormMuttActiveTraditional):
 
 
 class ClidApp(npy.NPSAppManaged):
-    def onStart(self):
+    """Class used by npyscreen to manage forms.
+
+       Attributes:
+            current_file(str):
+                file selected when in main_view(path)
+            settings(configobj.ConfigObj):
+                object used to read and write preferences
+    """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
         npy.setTheme(npy.Themes.ElegantTheme)
 
         self.current_file = None   # changed when a file is selected in main screen
+        self.settings = configobj.ConfigObj(os.path.expanduser('~/.clid.ini'))
+
+    def onStart(self):
         self.addForm("MAIN", ClidInterface)
         self.addForm("SETTINGS", pref.PreferencesView)
         self.addFormClass("EDIT", editmeta.EditMeta)   # addFormClass to create a new instance every time